requirements.txt
beautifulsoup4==4.6.0
Django==1.10.6
django-bootstrap3==8.2.3
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-filters==0.2.1
djangorestframework==3.6.2
et-xmlfile==1.0.1
fcm-django==0.2.8
gunicorn==19.7.1
jdcal==1.3
jsonfield==2.0.1
lxml==3.7.3
marshmallow==1.2.6
openpyxl==2.4.7
ordereddict==1.1
pep8==1.7.0
py==1.4.33
pyfcm==1.2.3
Pygments==2.2.0
PyMySQL==0.7.11
pytest==3.0.7
python-xbrl==1.1.1
requests==2.14.2
six==1.10.0
urllib3==1.21.1
